Moved filter Types under src/Filters and updated imports
